在运行库中实现Posix 的 opendir readdir 和closedir 三个主要接口。ls 命令使用他们为开发环境提供查看当前目录下文...

+++ include/kernel/ke_srv.h
@@ -41,7 +41,7 @@ struct ke_srv_info
 */
 struct sysreq_common
 {
-	unsigned long req_id;
+	unsigned int req_id;
 };
 
 bool ke_srv_register(const struct ke_srv_info *info);
